requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1

# Database
psycopg2-binary==2.9.9
alembic==1.13.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Email
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# File Processing
pandas==2.1.4
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.0
werkzeug==3.0.1
email-validator==2.1.0
